Very big commit, add support for PWM board.
Board type is now chosen in user_config.h, or can be overridden by the compiler. Hardware support is enabled/disabled based on this.
Board type can be read over the network, either by checking for the "PWM" object, or by reading the value of "GET BOARD TYPE".
For a quick sanity check, a PWM board has all 1s on the 7 seg on powerup, while an ADC board has all zeros.